+++ dio/CHANGELOG.md
@@ -5,7 +5,8 @@ See the [Migration Guide][] for the complete breaking changes list.**
 
 ## Unreleased
 
-*None.*
+- Remove `http` from `dev_dependencies`.
+- Only produce null response body when `ResponseType.json`.
 
 ## 5.2.1+1